添加头像

+++ dao/src/main/kotlin/com/qiqinote/dao/PictureDao.kt
@@ -7,7 +7,7 @@ import com.qiqinote.po.Picture
  * Created by vanki on 2018/1/23 14:04.
  */
 interface PictureDao {
-    fun insert(picture: Picture): Int
+    fun insert(picture: Picture): Long
 
     fun getById(id: Long): Picture?
